LayoutTests/fast/scrolling/scroll-animator-overlay-scrollbars-clicked-expected.txt

CONSOLE MESSAGE: line 14: MainFrameView: didAddVerticalScrollbar
CONSOLE MESSAGE: line 14: MainFrameView: didAddHorizontalScrollbar
CONSOLE MESSAGE: line 14: FrameView: didAddVerticalScrollbar
CONSOLE MESSAGE: line 14: FrameView: willRemoveVerticalScrollbar
CONSOLE MESSAGE: line 14: MainFrameView: mouseEnteredContentArea
CONSOLE MESSAGE: line 14: MainFrameView: mouseMovedInContentArea
CONSOLE MESSAGE: line 16: MainFrameView: mouseEnteredVerticalScrollbar
CONSOLE MESSAGE: line 16: FrameView: mouseEnteredContentArea
CONSOLE MESSAGE: line 16: MainFrameView: mouseMovedInContentArea
CONSOLE MESSAGE: line 17: MainFrameView: mouseIsDownInVerticalScrollbar
CONSOLE MESSAGE: line 18: MainFrameView: mouseIsUpInVerticalScrollbar
CONSOLE MESSAGE: line 19: MainFrameView: mouseExitedVerticalScrollbar
CONSOLE MESSAGE: line 19: FrameView: mouseExitedContentArea
CONSOLE MESSAGE: line 19: MainFrameView: mouseMovedInContentArea
PASS successfullyParsed is true

TEST COMPLETE
Test for https://bugs.webkit.org/show_bug.cgi?id=165056.
